front: move main actions from parent component in StdcmOpSchedule
