+++ front/src/applications/stdcm/components/StdcmForm/StdcmOrigin.tsx
@@ -5,24 +5,19 @@ import { useSelector } from 'react-redux';
 
 import { getTimesInfoFromDate } from 'applications/stdcm/utils';
 import OriginIcon from 'assets/pictures/mapMarkers/start.svg';
-import { useOsrdConfActions, useOsrdConfSelectors } from 'common/osrdContext';
-import type { StdcmConfSliceActions } from 'reducers/osrdconf/stdcmConf';
+import { useOsrdConfSelectors } from 'common/osrdContext';
 import type { StdcmConfSelectors } from 'reducers/osrdconf/stdcmConf/selectors';
-import { useAppDispatch } from 'store';
 
 import StdcmCard from './StdcmCard';
 import StdcmOperationalPoint from './StdcmOperationalPoint';
 import StdcmOpSchedule from './StdcmOpSchedule';
-import type { ArrivalTimeTypes, ScheduleConstraint, StdcmConfigCardProps } from '../../types';
+import type { StdcmConfigCardProps } from '../../types';
 
 const StdcmOrigin = ({ disabled = false }: StdcmConfigCardProps) => {
   const { t } = useTranslation('stdcm');
-  const dispatch = useAppDispatch();
 
   const { getStdcmOrigin } = useOsrdConfSelectors() as StdcmConfSelectors;
   const origin = useSelector(getStdcmOrigin);
-
-  const { updateStdcmPathStep } = useOsrdConfActions() as StdcmConfSliceActions;
 
   const { originArrival, originToleranceValues } = useMemo(
     () => ({
@@ -34,40 +29,6 @@ const StdcmOrigin = ({ disabled = false }: StdcmConfigCardProps) => {
     }),
     [origin]
   );
-
-  const onOriginArrivalChange = ({ date, hours, minutes }: ScheduleConstraint) => {
-    date.setHours(hours, minutes);
-    dispatch(
-      updateStdcmPathStep({
-        id: origin.id,
-        updates: { arrival: date },
-      })
-    );
-  };
-
-  const onOriginArrivalTypeChange = (arrivalType: ArrivalTimeTypes) => {
-    dispatch(
-      updateStdcmPathStep({
-        id: origin.id,
-        updates: { arrivalType },
-      })
-    );
-  };
-
-  const onOriginToleranceChange = ({
-    toleranceBefore,
-    toleranceAfter,
-  }: {
-    toleranceBefore: number;
-    toleranceAfter: number;
-  }) => {
-    dispatch(
-      updateStdcmPathStep({
-        id: origin.id,
-        updates: { tolerances: { before: toleranceBefore, after: toleranceAfter } },
-      })
-    );
-  };
 
   return (
     <StdcmCard
@@ -83,9 +44,7 @@ const StdcmOrigin = ({ disabled = false }: StdcmConfigCardProps) => {
         disabled={disabled}
       />
       <StdcmOpSchedule
-        onArrivalChange={onOriginArrivalChange}
-        onArrivalTypeChange={onOriginArrivalTypeChange}
-        onArrivalToleranceChange={onOriginToleranceChange}
+        pathStep={origin}
         opTimingData={originArrival}
         opToleranceValues={originToleranceValues}
         opScheduleTimeType={origin.arrivalType}
